Update master (#425)
* DOCKER: add Dockerfile

* DOCKER: update dockerfile

* DOCKER: update dockerfile

* DOCKER: path

* DOCKER: add cv docker file

* remove the tracking pipeline yml file

* README updates (#358)

* Updating environment.yml file in master (#323)

* readme updates

* mv media to scenarios folder

* fixes

* Update README.md

* simplification of language, removing redundancy

* added target audience section

* Update SETUP.md

* Update README.md

* Update environment.yml

* Update SETUP.md

* env-update (#359)

* Hyperdrive notebook updates (#356)

All tests are passing (except for unrelated AML deployment notebooks)

* transforms fix (#360)

* Updating environment.yml file in master (#323)

* fix for dataset transformations

* remove extra cython in conda

* pr comments'

* refactor to use transform in class param

* remove todo

* update to transformer

* added functionality to show transformations and updated notebook

* Update FAQ.md

* Adding contrib placeholder (#370)

* DOCKER: update readme

* adding missing lib dir

* add i3d

* Adding hard negative sampling notebook (#367)

* DOCKER: use create instead of update

* Add example gif to action recognition readme (#374)

* code clean up

* add i3d

* code clean up

* add action_recognition README content

* add instructions and headers

* fix conflicts

* DOCKER: remove base env bin path

* save/load detection code for deployment (#380)

* Updating environment.yml file in master (#323)

* save/load

* load/save

* load/save

* remove cython duplicate

* remove comment

* docstring

* tests for loading/saving

* label bug

* Syntax issues on lines 07 & 115 (#378)

* Updating environment.yml file in master (#323)

* update maximum time

* Restore example figures (#357)

* Staging (#365)

* README updates (#358)

* Updating environment.yml file in master (#323)

* readme updates

* mv media to scenarios folder

* fixes

* Update README.md

* simplification of language, removing redundancy

* added target audience section

* Update SETUP.md

* Update README.md

* Update environment.yml

* Update SETUP.md

* env-update (#359)

* Hyperdrive notebook updates (#356)

All tests are passing (except for unrelated AML deployment notebooks)

* Syntax issues on lines 07 & 115

* Update README.md

* Update environment.yml

* detection deploy model.py update (#381)

* Updating environment.yml file in master (#323)

* save/load

* load/save

* load/save

* remove cython duplicate

* remove comment

* docstring

* tests for loading/saving

* label bug

* initial notebook

* minor update to model.py

* revert hns nb

* rm nb

* ap at iou 0.5 (#385)

* Updating environment.yml file in master (#323)

* added ap_iou_05

* remove cython

* bug fix

* windows testing fix and other testing bugs (#383)

* Update azure-pipeline-windows-cpu.yml

* Update azure-pipeline-windows-gpu.yml

* Update test_integration_similarity_notebooks.py

* Update test_detection_notebooks.py

* 00 notebook (#386)

* Updating environment.yml file in master (#323)

* 00 notebook update

* remove extra dependency

* remove cython

* remove typo

* Update 10_image_annotation.ipynb

Added link to Azure annotation tool

* hyperparam remark (#388)

* Updating environment.yml file in master (#323)

* add concluding remark to build the store for the hyperdrive notebook.

* Update environment.yml

* initial commit

* Count based OD evaluation (#415)

* Update README.md

* Update README.md

* Notebook for deploying a detection model (#382)

* Updating environment.yml file in master (#323)

* save/load

* load/save

* load/save

* remove cython duplicate

* remove comment

* docstring

* tests for loading/saving

* label bug

* initial notebook

* minor update to model.py

* revert hns nb

* rm nb

* deployment notebook

* update to deployment notebook

* update to notebok

* notebook fix

* minor fixes to notebook + testS

* fix pr

* Add Mask R-CNN (#389)

* increased lr for head by 10x (#421)

* fixes from bugbash (#422)

* Update SETUP.md

* Update SETUP.md

* Update README.md

* fix automatic merge error

+++ utils_cv/detection/references/coco_eval.py
@@ -114,7 +114,9 @@ class CocoEvaluator(object):
 
             rles = [
                 mask_util.encode(
-                    np.array(mask[0, :, :, np.newaxis], order="F")
+                    # Change according to the issue related to mask:
+                    #     https://github.com/pytorch/vision/issues/1355#issuecomment-544951911
+                    np.array(mask[0, :, :, np.newaxis], dtype=np.uint8, order="F")
                 )[0]
                 for mask in masks
             ]
